Credits and Warps
Started adding the /balance command and /warp command. Balance is working but does not currently save balances if the server restarts. This will be added. Warps are not yet functional. Players get credits for killing zombies. $5 for a regular zombie and $50 for a mega zombie. This can be altered in the config. I also added a money effect when you kill a zombie or check your /balance. CHA CHING