Player Chat Suggestions
Added a feature where if a player says, in global or local chat, "how do i buy a vehicle" or "how do I know the price of an item" a response from the server will appear in their chat and give them the commands related to what they asked. Also added a config option to disable this. May expand on this later.

Added a check to the load function to check for players currently on the server and load them in dynamically so that the plugin can be reloaded without error spamming or having to restart the server. Noice.